feat(component): setting up the people page
I've renamed  the old peoples component, and created a new 'People' component for the new UI change

2176
