+++ apps/user-profile/src/components/modules/UserProfile/People[OLD]/StyledPeople.js
@@ -1,4 +1,4 @@
-import styled, { css } from "styled-components";
+import styled, { css } from 'styled-components';
 
 export const PeopleContainer = styled.div`
   /* width: 100%; */
